Add unbalanced returns tests.
