+++ tests/parser/syntax/test_bytes.py
@@ -60,6 +60,7 @@ def foo(x: int128) -> bytes[75]:
 def foo() -> bytes[10]:
     x: bytes[10] = '0x1234567890123456789012345678901234567890'
     x = 0x1234567890123456789012345678901234567890
+    return x
     """,
     """
 @public
